initial repairs
