+++ src/simplnx/Parameters/FileSystemPathParameter.cpp
@@ -2,168 +2,20 @@
 
 #include "simplnx/Common/Any.hpp"
 #include "simplnx/Common/StringLiteral.hpp"
+#include "simplnx/Utilities/FileUtilities.hpp"
 #include "simplnx/Utilities/StringUtilities.hpp"
 
 #include <fmt/core.h>
 
 #include <nlohmann/json.hpp>
 
-#include <cctype>
 #include <filesystem>
-#include <iostream>
 #include <stdexcept>
-
-#ifdef _WIN32
-#include <io.h>
-#define FSPP_ACCESS_FUNC_NAME _access
-#else
-#include <unistd.h>
-#define FSPP_ACCESS_FUNC_NAME access
-#endif
 
 namespace fs = std::filesystem;
 
 using namespace nx::core;
 
-namespace
-{
-#ifdef _WIN32
-constexpr int k_CheckWritable = 2;
-#else
-constexpr int k_CheckWritable = W_OK;
-#endif
-
-constexpr StringLiteral k_PathKey = "path";
-constexpr int k_HasAccess = 0;
-
-//-----------------------------------------------------------------------------
-bool HasWriteAccess(const std::string& path)
-{
-  return FSPP_ACCESS_FUNC_NAME(path.c_str(), k_CheckWritable) == k_HasAccess;
-}
-
-//-----------------------------------------------------------------------------
-Result<> ValidateInputFile(const FileSystemPathParameter::ValueType& path)
-{
-  if(!fs::exists(path))
-  {
-    return MakeErrorResult(-2, fmt::format("File System Path '{}' does not exist", path.string()));
-  }
-
-  if(!fs::is_regular_file(path))
-  {
-    return MakeErrorResult(-3, fmt::format("File System Path '{}' is not a file", path.string()));
-  }
-  return {};
-}
-
-//-----------------------------------------------------------------------------
-Result<> ValidateInputDir(const FileSystemPathParameter::ValueType& path)
-{
-  if(!fs::exists(path))
-  {
-    return MakeErrorResult(-4, fmt::format("File System Path '{}' does not exist", path.string()));
-  }
-  if(!fs::is_directory(path))
-  {
-    return MakeErrorResult(-5, fmt::format("File System Path '{}' is not a file", path.string()));
-  }
-  return {};
-}
-//-----------------------------------------------------------------------------
-Result<> ValidateDirectoryWritePermission(const FileSystemPathParameter::ValueType& path, bool isFile)
-{
-  if(path.empty())
-  {
-    return MakeErrorResult(-16, "ValidateDirectoryWritePermission() error: given path was empty.");
-  }
-
-  auto checkedPath = path;
-  if(isFile)
-  {
-    checkedPath = checkedPath.parent_path();
-  }
-  // We now have the parent directory. Let us see if *any* part of the path exists
-
-  // If the path is relative, then make it absolute
-  if(!checkedPath.is_absolute())
-  {
-    try
-    {
-      checkedPath = fs::absolute(checkedPath);
-    } catch(const std::filesystem::filesystem_error& error)
-    {
-      return MakeErrorResult(-15, fmt::format("ValidateDirectoryWritePermission() threw an error: '{}'", error.what()));
-    }
-  }
-
-  auto rootPath = checkedPath.root_path();
-
-  // The idea here is to start walking up from the deepest directory and hopefully
-  // find an existing directory. If we get to the top if the path and we are still
-  // empty then:
-  //  On unix based systems not sure if it would happen. Even if the user set a path
-  // to another drive that didn't exist, at some point you hit the '/' and then you
-  // can try to create the directories.
-  //  On Windows the user put in a bogus drive letter which is just a hard failure
-  // because we can't make up a new drive letter.
-  while(!fs::exists(checkedPath) && checkedPath != rootPath)
-  {
-    checkedPath = checkedPath.parent_path();
-  }
-
-  if(checkedPath.empty())
-  {
-    return MakeErrorResult(-19, "ValidateDirectoryWritePermission() resolved path was empty");
-  }
-
-  if(!fs::exists(checkedPath))
-  {
-    return MakeErrorResult(-11, fmt::format("ValidateDirectoryWritePermission() error: The drive does not exist on this system: '{}'", checkedPath.string()));
-  }
-
-  // We should be at the top of the tree with an existing directory.
-  if(HasWriteAccess(checkedPath.string()))
-  {
-    return {};
-  }
-  return MakeErrorResult(-8, fmt::format("User does not have write permissions to path '{}'", path.string()));
-}
-
-//-----------------------------------------------------------------------------
-Result<> ValidateOutputFile(const FileSystemPathParameter::ValueType& path)
-{
-  auto result = ValidateDirectoryWritePermission(path, true);
-  if(result.invalid())
-  {
-    return result;
-  }
-  if(!fs::exists(path))
-  {
-    return MakeWarningVoidResult(-6, fmt::format("File System Path '{}' does not exist. It will be created during execution.", path.string()));
-  }
-  return {};
-}
-
-//-----------------------------------------------------------------------------
-Result<> ValidateOutputDir(const FileSystemPathParameter::ValueType& path)
-{
-  auto result = ValidateDirectoryWritePermission(path, false);
-  if(result.invalid())
-  {
-    return result;
-  }
-  if(!fs::exists(path))
-  {
-    return MakeWarningVoidResult(-7, fmt::format("File System Path '{}' does not exist. It will be created during execution.", path.string()));
-  }
-  return {};
-}
-
-} // namespace
-
-namespace nx::core
-{
 //-----------------------------------------------------------------------------
 FileSystemPathParameter::FileSystemPathParameter(const std::string& name, const std::string& humanName, const std::string& helpText, const ValueType& defaultValue,
                                                  const ExtensionsType& extensionsType, PathType pathType, bool acceptAllExtensions)
@@ -182,11 +34,11 @@ FileSystemPathParameter::FileSystemPathParameter(const std::string& name, const 
     }
     if(ext.at(0) != '.')
     {
-      validatedExtensions.insert('.' + nx::core::StringUtilities::toLower(ext));
+      validatedExtensions.insert('.' + StringUtilities::toLower(ext));
     }
     else
     {
-      validatedExtensions.insert(nx::core::StringUtilities::toLower(ext));
+      validatedExtensions.insert(StringUtilities::toLower(ext));
     }
   }
   m_AvailableExtensions = validatedExtensions;
@@ -276,16 +128,16 @@ Result<> FileSystemPathParameter::validatePath(const ValueType& path) const
 
   if(path.empty())
   {
-    return nx::core::MakeErrorResult(-3001, fmt::format("{} File System Path must not be empty", prefix));
+    return MakeErrorResult(-3001, fmt::format("{} File System Path must not be empty", prefix));
   }
 
-  if(!m_acceptAllExtensions && (m_PathType == nx::core::FileSystemPathParameter::PathType::InputFile || m_PathType == nx::core::FileSystemPathParameter::PathType::OutputFile))
+  if(!m_acceptAllExtensions && (m_PathType == FileSystemPathParameter::PathType::InputFile || m_PathType == FileSystemPathParameter::PathType::OutputFile))
   {
     if(!path.has_extension())
     {
       return {nonstd::make_unexpected(std::vector<Error>{{-3002, fmt::format("{} File System Path must include a file extension", prefix)}})};
     }
-    std::string lowerExtension = nx::core::StringUtilities::toLower(path.extension().string());
+    std::string lowerExtension = StringUtilities::toLower(path.extension().string());
     if(path.has_extension() && !m_AvailableExtensions.empty() && m_AvailableExtensions.find(lowerExtension) == m_AvailableExtensions.end())
     {
       return {nonstd::make_unexpected(std::vector<Error>{{-3003, fmt::format("{} File extension '{}' is not a valid file extension", prefix, path.extension().string())}})};
@@ -294,16 +146,15 @@ Result<> FileSystemPathParameter::validatePath(const ValueType& path) const
 
   switch(m_PathType)
   {
-  case nx::core::FileSystemPathParameter::PathType::InputFile:
-    return ValidateInputFile(path);
-  case nx::core::FileSystemPathParameter::PathType::InputDir:
-    return ValidateInputDir(path);
-  case nx::core::FileSystemPathParameter::PathType::OutputFile:
-    return ValidateOutputFile(path);
-  case nx::core::FileSystemPathParameter::PathType::OutputDir:
-    return ValidateOutputDir(path);
+  case FileSystemPathParameter::PathType::InputFile:
+    return FileUtilities::ValidateInputFile(path);
+  case FileSystemPathParameter::PathType::InputDir:
+    return FileUtilities::ValidateInputDir(path);
+  case FileSystemPathParameter::PathType::OutputFile:
+    return FileUtilities::ValidateOutputFile(path);
+  case FileSystemPathParameter::PathType::OutputDir:
+    return FileUtilities::ValidateOutputDir(path);
   }
 
   return {};
 }
-} // namespace nx::core
